Error nicely if pip clobbers a conda-installed file (#35)
If pip and conda are mixed, sometimes pip will uninstall packages that
are managed by conda. This *can* lead to an inconsistent environment,
and is generally frowned upon. Instead of trying to handle these cases,
we collect packages that have been broken by pip and report a nice error
message on failure.