Allow specifying destination prefix before packing
This adds a new option `--dest-prefix` that allows specifying the
destination prefix before packing. In this case the prefixes are
rewritten before packing the environment, and the `conda-unpack` script
won't be generated.

This is useful for cases where you can't run the `conda-unpack` script
but know the destination path beforehand.