Skip fixing prefix in python exe in conda-unpack (#50)
On Linux rewriting a running executable file is forbidden. Python is
required for running `conda-unpack`, and *some* versions of python have
an absolute path in them - this can cause problems when the packaged
python is the one running the `conda-unpack` script.

Since the absolute path embedded in the python binary is not important
for normal use, we skip rewriting it when run from `conda-unpack` -
avoiding the issue entirely. Note that we still rewrite the prefix when
using a specified destination prefix.