Handle noarch-python packages missing pyc files (#64)
Some packages have files that will fail to bytecode compile on older
versions of python (e.g. optional py3-only features). When these files
are made noarch-python packages, some of them will fail to bytecode
compile (and thus won't exist), but conda still marks them as present
in the paths.json file. To correct for this, we now check for the
presence of every pyc file before adding it when collecting files. The
`jinja2` package exhibits this behavior, and valid packaging of an
environment with this package is used as a test.