Windows test fixes
Conda changed its activation scripts, which broke our windows tests.
This fixes those, as well as a lingering issue with pip/conda
compatibility checking.

In the case of packaging an environment that contains conda with a
specified destination prefix (``--dest-prefix``) on windows, things
still are slightly off. When conda is present and the destination is
specified we leave conda's activation scripts as is (don't provide our
own). Conda's new activation scripts for windows no longer work without
a base conda install AFAICT. Since this use case is small, the user
likely already has conda, and I don't see a good work around for this,
we fix the test and move on for now.